# tests/integration/fastapi_tests/snapshots/demo3_openapi.json
{
  "components": {
    "schemas": {
      "AppendOp": {
        "additionalProperties": true,
        "description": "Appends a value to an array.",
        "properties": {
          "op": {
            "const": "append",
            "default": "append",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "items": {},
              "type": "array"
            }
          },
          "value": {
            "title": "Value"
          }
        },
        "required": ["op", "path", "value"],
        "title": "Append operation",
        "type": "object"
      },
      "ConfigId": {
        "enum": ["service"],
        "title": "ConfigId",
        "type": "string"
      },
      "EnsureObjectOp": {
        "additionalProperties": true,
        "description": "Ensures the target path resolves to an object, creating one if missing.",
        "properties": {
          "op": {
            "const": "ensure_object",
            "default": "ensure_object",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "required": ["op", "path"],
        "title": "Ensure object operation",
        "type": "object"
      },
      "ExtendOp": {
        "additionalProperties": true,
        "description": "Extends an array with a list of values.",
        "properties": {
          "op": {
            "const": "extend",
            "default": "extend",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "items": {},
              "type": "array"
            }
          },
          "values": {
            "items": {},
            "title": "Values",
            "type": "array"
          }
        },
        "required": ["op", "path", "values"],
        "title": "Extend operation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncrementOp": {
        "additionalProperties": true,
        "description": "Increments a numeric field by a value.",
        "properties": {
          "op": {
            "const": "increment",
            "default": "increment",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "type": "number"
            }
          },
          "value": {
            "gt": 0,
            "multiple_of": 5,
            "title": "Value",
            "type": "number"
          }
        },
        "required": ["op", "path", "value"],
        "title": "Increment operation",
        "type": "object"
      },
      "RemoveNumberOp": {
        "additionalProperties": true,
        "description": "Removes a numeric value at a path.",
        "properties": {
          "op": {
            "const": "remove_number",
            "default": "remove_number",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "type": "number"
            }
          }
        },
        "required": ["op", "path"],
        "title": "Remove number operation",
        "type": "object"
      },
      "ServiceConfigPatchOperation": {
        "description": "Discriminated union of patch operations for ServiceConfig.",
        "discriminator": {
          "mapping": {
            "append": "#/components/schemas/AppendOp",
            "ensure_object": "#/components/schemas/EnsureObjectOp",
            "extend": "#/components/schemas/ExtendOp",
            "increment": "#/components/schemas/IncrementOp",
            "remove_number": "#/components/schemas/RemoveNumberOp",
            "set_message": "#/components/schemas/SetMessageOp",
            "swap": "#/components/schemas/SwapOp",
            "toggle": "#/components/schemas/ToggleBoolOp"
          },
          "propertyName": "op"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AppendOp"
          },
          {
            "$ref": "#/components/schemas/EnsureObjectOp"
          },
          {
            "$ref": "#/components/schemas/ExtendOp"
          },
          {
            "$ref": "#/components/schemas/IncrementOp"
          },
          {
            "$ref": "#/components/schemas/RemoveNumberOp"
          },
          {
            "$ref": "#/components/schemas/SetMessageOp"
          },
          {
            "$ref": "#/components/schemas/SwapOp"
          },
          {
            "$ref": "#/components/schemas/ToggleBoolOp"
          }
        ],
        "title": "ServiceConfig Patch Operation"
      },
      "ServiceConfigPatchRequest": {
        "description": "Array of patch operations for ServiceConfig.",
        "items": {
          "$ref": "#/components/schemas/ServiceConfigPatchOperation"
        },
        "title": "ServiceConfig Patch Request",
        "type": "array"
      },
      "SetMessageOp": {
        "additionalProperties": true,
        "description": "Sets a string field to text or null.",
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "op": {
            "const": "set_message",
            "default": "set_message",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            }
          }
        },
        "required": ["message", "op", "path"],
        "title": "Set message operation",
        "type": "object"
      },
      "SwapOp": {
        "additionalProperties": true,
        "description": "Swaps the values at paths a and b. Paths a and b may not be proper prefixes of each other.",
        "properties": {
          "a": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "A",
            "type": "string",
            "x-pointer-type-schema": {}
          },
          "b": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "B",
            "type": "string",
            "x-pointer-type-schema": {}
          },
          "op": {
            "const": "swap",
            "default": "swap",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          }
        },
        "required": ["a", "b", "op"],
        "title": "Swap operation",
        "type": "object"
      },
      "ToggleBoolOp": {
        "additionalProperties": true,
        "description": "Toggles a boolean value at a path.",
        "properties": {
          "op": {
            "const": "toggle",
            "default": "toggle",
            "description": "The operation to perform.",
            "title": "Op",
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer (RFC 6901) string",
            "format": "json-pointer",
            "title": "Path",
            "type": "string",
            "x-pointer-type-schema": {
              "type": "boolean"
            }
          }
        },
        "required": ["op", "path"],
        "title": "Toggle operation",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Plain JSON patching for service configs using `JsonPatchFor[Name, Registry]`.",
    "title": "Demo 3: Control plane configs",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/configs/{config_id}": {
      "get": {
        "description": "Fetch a service config by id.",
        "operationId": "get_config_endpoint_configs__config_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConfigId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Config Endpoint Configs  Config Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a service config",
        "tags": ["configs"]
      },
      "patch": {
        "description": "Apply standard RFC 6902 ops plus custom ops to a service config.",
        "operationId": "patch_config_configs__config_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "config_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConfigId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "examples": {
                "maintenance-note": {
                  "summary": "set or clear the service_name message",
                  "value": [
                    {
                      "message": "Atlas maintenance in progress",
                      "op": "set_message",
                      "path": "/service_name"
                    }
                  ]
                },
                "oops-expected": {
                  "summary": "type-gated remove (expected failure)",
                  "value": [
                    {
                      "op": "ensure_object",
                      "path": "/features"
                    },
                    {
                      "op": "remove_number",
                      "path": "/service_name"
                    }
                  ]
                },
                "rocket-boost": {
                  "summary": "bump max_users and toggle chat",
                  "value": [
                    {
                      "op": "increment",
                      "path": "/limits/max_users",
                      "value": 50
                    },
                    {
                      "op": "toggle",
                      "path": "/features/chat"
                    }
                  ]
                },
                "shuffle-switch": {
                  "summary": "swap and then toggle",
                  "value": [
                    {
                      "a": "/service_name",
                      "b": "/features/chat",
                      "op": "swap"
                    },
                    {
                      "op": "toggle",
                      "path": "/features/chat"
                    }
                  ]
                },
                "tag-and-seal": {
                  "summary": "ensure features and append a tag",
                  "value": [
                    {
                      "op": "ensure_object",
                      "path": "/features"
                    },
                    {
                      "op": "append",
                      "path": "/tags",
                      "value": "beta"
                    }
                  ]
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ServiceConfigPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Patch Config Configs  Config Id  Patch"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "properties": {
                            "cause_type": {
                              "type": ["string", "null"]
                            },
                            "index": {
                              "type": "integer"
                            },
                            "message": {
                              "type": "string"
                            },
                            "op": {
                              "type": "object"
                            }
                          },
                          "required": ["index", "op", "message"],
                          "type": "object"
                        }
                      ]
                    }
                  },
                  "required": ["detail"],
                  "type": "object"
                }
              }
            },
            "description": "Patch cannot be applied to current resource state"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "properties": {
                            "cause_type": {
                              "type": ["string", "null"]
                            },
                            "index": {
                              "type": "integer"
                            },
                            "message": {
                              "type": "string"
                            },
                            "op": {
                              "type": "object"
                            }
                          },
                          "required": ["index", "op", "message"],
                          "type": "object"
                        }
                      ]
                    }
                  },
                  "required": ["detail"],
                  "type": "object"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "properties": {
                        "detail": {
                          "oneOf": [
                            {
                              "type": "string"
                            },
                            {
                              "properties": {
                                "cause_type": {
                                  "type": ["string", "null"]
                                },
                                "index": {
                                  "type": "integer"
                                },
                                "message": {
                                  "type": "string"
                                },
                                "op": {
                                  "type": "object"
                                }
                              },
                              "required": ["index", "op", "message"],
                              "type": "object"
                            }
                          ]
                        }
                      },
                      "required": ["detail"],
                      "type": "object"
                    },
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    }
                  ]
                }
              }
            },
            "description": "Request validation or patch document validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "properties": {
                            "cause_type": {
                              "type": ["string", "null"]
                            },
                            "index": {
                              "type": "integer"
                            },
                            "message": {
                              "type": "string"
                            },
                            "op": {
                              "type": "object"
                            }
                          },
                          "required": ["index", "op", "message"],
                          "type": "object"
                        }
                      ]
                    }
                  },
                  "required": ["detail"],
                  "type": "object"
                }
              }
            },
            "description": "Patch execution error"
          }
        },
        "summary": "Patch a service config",
        "tags": ["configs"]
      }
    }
  }
}